Доработана реализация многопоточных фильтров. Корректная обработка CSRF-токенов. OAuth2 с помощью google id token.
weekend

weekend

weekend

+++ foodsaver/src/main/kotlin/com/foodback/dto/request/cart/CartRequestModel.kt
@@ -14,5 +14,5 @@ data class CartRequestModel(
     val productId: UUID,
 
     @field:Min(value = 1, message = "Min quantity 1")
-    val quantity: Int = 1
+    val quantity: Long = 1
 )
